BLE final updates: temporarily using middleware from STM32 firmware, but still communicating using my custom SPI.
Testing done and successful.

+++ Drivers/Inc/Timer.h
@@ -173,7 +173,7 @@ FUNCTIONS PROTOTYPES
 ****************************************************************************/
 
 /* Create delay in milliseconds */
-void Delay_ms(uint32_t ms);
+void Delay_ms(float ms);
 
 /* Initialize CPU SysTick and General-Purpose Timer peripheral */
 void Timer_Init();
